// src/app/enums/window-types.enum.ts
export enum WindowTypes {
  LAUNCHPAD = 'LAUNCHPAD',
  GALLERY = 'GALLERY',
  SOUNDCLOUD_PLAYER = 'SOUNDCLOUD_PLAYER',
  SETTINGS = 'SETTINGS',
  ABOUT = 'ABOUT',
  NOTEPAD = 'NOTEPAD',
  TETRIS = 'TETRIS',
  STYLE_GUIDE = 'STYLE_GUIDE',
  LOGIN = 'LOGIN',
  MINT_CTA = 'MINT_CTA',
  CALCULATOR = 'CALCULATOR',
  MEMORY_MATCH = 'MEMORY_MATCH',
  SNAKE = 'SNAKE',
  POLL = 'POLL',
  SLOT_MACHINE = 'SLOT_MACHINE',
  PONG = 'PONG',
  MINESWEEPER = 'MINESWEEPER',
  GAME_2048 = 'GAME_2048',
  BREAKOUT = 'BREAKOUT',
  FLAPPY_RAVEN = 'FLAPPY_RAVEN',
  NFT_SHOWCASE = 'NFT_SHOWCASE',
  LEADERBOARD = 'LEADERBOARD',
  RARITY_CHECKER = 'RARITY_CHECKER',
  ROADMAP = 'ROADMAP',
  MONO = 'MONO',
  MORSE_CODE = 'MORSE_CODE',
  TYPING_TEST = 'TYPING_TEST',
  WORLD_CLOCK = 'WORLD_CLOCK',
  CODE_PLAYGROUND = 'CODE_PLAYGROUND',
  PD_RADIO = 'PD_RADIO',
  MARKDOWN_WIKI = 'MARKDOWN_WIKI',
  NFT_GALLERY = 'NFT_GALLERY',
}